adding support for other widths in ocl driver

+++ soda_codes/harris_mini_opt/our_code/harris_mini_opt_host.cpp
@@ -31,10 +31,9 @@ int main(int argc, char **argv) {
 
   std::ofstream input_img_update_0_read("img_update_0_read.csv");
   for (int i = 0; i < img_update_0_read_DATA_SIZE; i++) {
-    // TODO: Add support for other widths
-    uint16_t val = (i % 256);
+    uint32_t val = (i % 256);
     input_img_update_0_read << val << std::endl;
-    ((uint16_t*) (img_update_0_read.data()))[i] = val;
+    ((uint32_t*) (img_update_0_read.data()))[i] = val;
   }
 
   input_img_update_0_read.close();
